moved request details to request folder
